+++ src/day24_Arrays/UniqueWords2.java
@@ -3,11 +3,17 @@ package day24_Arrays;
 public class UniqueWords2 {
 
     public static void main(String[] args) {
+        uniqueWordsInArray();
+
+
+    }
+
+    private static void uniqueWordsInArray(){
         String[] words = {"C#", "C#", "Java", "Python", "Python", "C++", "Ruby"};
 
         for(String each: words){
             int count = 0;
-            for( String b  : words  ){ // b represents each elemnts in the array
+            for( String b  : words  ){ // b represents each elements in the array
                 if(each.equals(b)){
                     count++;
                 }
@@ -18,12 +24,6 @@ public class UniqueWords2 {
             }
 
         }
-
-
-
-
-
-
 
     }
